When chaos reigned in lines of code, A daunting task, a heavy load, The state machine needed a brand-new frame, Porting state.c was the name of the game. Confusion loomed, bugs ran amok, Progress slowed, and we felt stuck. But fear not, for in walked Liam Merino, Our coding hero, our project’s savior-ino! With steady hands and a sharp, keen eye, He tackled the task, gave it a try. Through functions deep and logic tight, He helped jshen3 rewrote state.c till it ran just right. Debugging lines, they stayed up late, Ensuring transitions aligned with fate. No variable orphaned, no state left behind, A coder of vision, resourceful and kind. The syntax shone, the errors were slain, In newState.c, we saw our gain. Their work was a masterpiece fine, Efficiency and elegance, perfectly aligned. So here's to Liam, the clutch MVP, Who saved the day with pure mastery. Our gratitude flows in every commit, For the legend of Merino will never quit.
